add compatibility with doctrine/dbal 3
