+++ tests/DAMA/DoctrineTestBundle/Doctrine/DBAL/MockDriver.php
@@ -2,62 +2,14 @@
 
 namespace Tests\DAMA\DoctrineTestBundle\Doctrine\DBAL;
 
-use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\Driver;
-use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Schema\AbstractSchemaManager;
-use PHPUnit\Framework\MockObject\Generator;
-
-class MockDriver implements Driver
-{
-    private function getMock(string $class)
+if (interface_exists(\Doctrine\DBAL\Driver\ExceptionConverterDriver::class)) {
+    // dbal v2
+    class MockDriver extends MockDriverV2
     {
-        return (new Generator())->getMock(
-            $class,
-            [],
-            [],
-            '',
-            false
-        );
     }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function connect(array $params, $username = null, $password = null, array $driverOptions = []): Driver\Connection
+} else {
+    // dbal v3
+    class MockDriver extends MockDriverV3
     {
-        return $this->getMock(Driver\Connection::class);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabasePlatform(): AbstractPlatform
-    {
-        return $this->getMock(AbstractPlatform::class);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getSchemaManager(Connection $conn): AbstractSchemaManager
-    {
-        return $this->getMock(AbstractSchemaManager::class);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName(): string
-    {
-        return 'mock';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabase(Connection $conn): string
-    {
-        return 'mock';
     }
 }
