--- tests/Functional/app/parameters.yml
@@ -1,5 +1,0 @@
-parameters:
-    db_host: 127.0.0.1
-    db_name: doctrine_test_bundle
-    db_user: root
-    db_password: ~
